requirements.txt
Django==4.1.4
celery==5.2.7
crispy-bootstrap5==0.7
django-crispy-forms==1.14.0
djangorestframework==3.14.0
django-environ==0.9.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
whitenoise==6.3.0
